refactor: adopt shared hooks in converted tables
Extracted reusable pagination and active-state logic into Hooks/tables
and ensured all tables already using data-view consume these hooks
instead of hand-rolling the same patterns.
